Pandora::Utils classes extraction
